Update dependencies, test configuration
Remove invoke for increased compatibility, add setup.cfg for test and
linting configuration, add isort and yapf for increased control over
styles, update Travis to test multiple Python versions.
