+++ Pipfile
@@ -5,25 +5,25 @@ name = "pypi"
 
 [packages]
 airtable-python-wrapper = "*"
-betamax = "*"
-betamax-serializers = "*"
 boto3 = "*"
-flake8 = "*"
-freezegun = "*"
 inflector = "*"
-invoke = "*"
 jinja2 = "*"
 legistar = {git = "https://github.com/opencivicdata/python-legistar-scraper" } 
-pep8 = "*"
-pytest = "*"
 python-dateutil = "*"
 pytz = "*"
 requests = "*"
-scrapy = "*"
-scrapy-sentry ="*"
 usaddress = "*"
 ics = "*"
 azure-storage-blob = "*"
+twisted = ">=18.9.0"
+scrapy = "*"
+scrapy-sentry = "*"
 
-[requires]
-python_version = "3.6"
+[dev-packages]
+freezegun = "*"
+pytest = "*"
+"flake8" = "*"
+betamax = "*"
+betamax-serializers = "*"
+isort = "*"
+yapf = "*"
