+++ requirements.txt
@@ -2,64 +2,66 @@
 airtable-python-wrapper==0.11.3.post1
 arrow==0.4.2
 asn1crypto==0.24.0
-atomicwrites==1.2.1; python_version >= '2.7'
 attrs==18.2.0
 automat==0.7.0
-azure-storage-blob==1.3.1
-betamax-serializers==0.2.0
-betamax==0.8.1
-boto3==1.8.8
-botocore==1.11.8
-certifi==2018.8.24
+azure-common==1.1.16
+azure-storage-blob==1.4.0
+azure-storage-common==1.4.0
+boto3==1.9.46
+botocore==1.12.46
+certifi==2018.10.15
 cffi==1.11.5
 chardet==3.0.4
 constantly==15.1.0
-cryptography==2.3.1
-cssselect==1.0.3; python_version >= '2.7'
+cryptography==2.4.1
+cssselect==1.0.3
 docutils==0.14
-flake8==3.5.0
-freezegun==0.3.10
-future==0.16.0
+future==0.17.1
 git+https://github.com/opencivicdata/python-legistar-scraper@912a5ac244f7c1ed593eeb4578e9f24a46dee7a2#egg=legistar
 hyperlink==18.0.0
 ics==0.4
 idna==2.7
 incremental==17.5.0
 inflector==2.0.12
-invoke==1.1.1
 jinja2==2.10
 jmespath==0.9.3
-lxml==4.2.4
-markupsafe==1.0
-mccabe==0.6.1
-more-itertools==4.3.0
-parsel==1.5.0; python_version >= '2.7'
-pep8==1.7.1
-pluggy==0.7.1; python_version != '3.2.*'
+lxml==4.2.5
+markupsafe==1.1.0
+parsel==1.5.1
 probableparsing==0.0.1
-py==1.6.0; python_version != '3.2.*'
 pyasn1-modules==0.2.2
 pyasn1==0.4.4
-pycodestyle==2.3.1
-pycparser==2.18
+pycparser==2.19
 pydispatcher==2.0.5
-pyflakes==1.6.0
-pyhamcrest==1.9.0; python_version >= '2.7'
+pyhamcrest==1.9.0
 pyopenssl==18.0.0
-pytest==3.8.0
 python-crfsuite==0.9.6
-python-dateutil==2.7.3
-pytz==2018.5
+python-dateutil==2.7.5
+pytz==2018.7
 queuelib==1.5.0
 raven==6.9.0
-requests==2.19.1
+requests==2.20.1
 s3transfer==0.1.13
 scrapy-sentry==0.9.0
 scrapy==1.5.1
 service-identity==17.0.0
 six==1.11.0
-twisted==18.7.0
-urllib3==1.23
+twisted==18.9.0
+urllib3==1.24.1 ; python_version >= '3.4'
 usaddress==0.5.10
 w3lib==1.19.0
-zope.interface==4.5.0; python_version >= '2.7'
+zope-interface==4.6.0
+atomicwrites==1.2.1
+betamax-serializers==0.2.0
+betamax==0.8.1
+flake8==3.6.0
+freezegun==0.3.11
+isort==4.3.4
+mccabe==0.6.1
+more-itertools==4.3.0
+pluggy==0.8.0
+py==1.7.0
+pycodestyle==2.4.0
+pyflakes==2.0.0
+pytest==4.0.0
+yapf==0.24.0
