+++ setup.cfg
@@ -1,0 +1,50 @@
+[flake8]
+exclude =
+    .git,
+    .venv,
+    venv,
+    docs,
+    */__pycache__/*,
+    tests/files/*,
+    tests/cassettes/*,
+max-line-length = 100
+
+[tool:pytest]
+python_files = tests.py test_*.py *_tests.py
+
+[isort]
+default_section = THIRDPARTY
+known_first_party = city_scrapers
+skip_glob =
+    */.venv/*
+sections = FUTURE,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER
+multi_line_output = 5
+line_length = 100
+
+[yapf]
+align_closing_bracket_with_visual_indent = True
+allow_multiline_lambdas = True
+allow_multiline_dictionary_keys = False
+allow_split_before_dict_value = False
+blank_line_before_class_docstring = False
+blank_line_before_nested_class_or_def = False
+coalesce_brackets = True
+column_limit = 100
+continuation_indent_width = 4
+dedent_closing_brackets = True
+each_dict_entry_on_separate_line = True
+indent_dictionary_value = True
+indent_width = 4
+join_multiple_lines = True
+spaces_around_default_or_named_assign = False
+spaces_before_comment = 2
+space_between_ending_comma_and_closing_bracket = False
+split_arguments_when_comma_terminated = True
+split_before_bitwise_operator = True
+split_before_dict_set_generator = True
+split_before_expression_after_opening_paren = False
+split_before_first_argument = False
+split_before_logical_operator = True
+split_before_named_assigns = True
+split_complex_comprehension = True
+use_tabs = False
