+++ travis/validate_spiders.sh
@@ -23,4 +23,4 @@ git diff --name-only --diff-filter=AM $TRAVIS_COMMIT_RANGE| \
 if [ $? -ne 0 ]; then exit 1; fi
 
 # Validate saved output
-find travis -name "*.json" | xargs -I{} invoke validate-spider {}
+find travis -name "*.json" | xargs -I{} python scripts/validate_spider.py {}
